+++ frontend/views/containers/chatroom/video-viewer/VideoPlayer.vue
@@ -11,7 +11,6 @@
 </template>
 
 <script>
-import { checkBrowserVideoMimeTypeSupport } from '../file-attachment/attachment-utils'
 import Plyr from 'plyr'
 
 export default {
@@ -45,8 +44,7 @@ export default {
         player: null
       },
       ephemeral: {
-        isTallerThanWider: false,
-        browserCanPlay: false
+        isTallerThanWider: false
       }
     }
   },
@@ -102,14 +100,6 @@ export default {
     isPlaying () {
       return this.config.player.playing
     },
-    checkBrowserMimeTypeSupport () {
-      if (this.mimeType) {
-        this.ephemeral.browserCanPlay = checkBrowserVideoMimeTypeSupport(this.mimeType)
-      }
-    }
-  },
-  created () {
-    this.checkBrowserMimeTypeSupport()
   },
   mounted () {
     this.initPlayer()
