+++ frontend/views/utils/filters.js
@@ -43,6 +43,18 @@ export const formatBytesDecimal = (bytes: number, decimals: number = 2): string 
   return `${formattedValue} ${sizes[i]}`
 }
 
+export function checkBrowserVideoMimeTypeSupport (mimeType = ''): boolean {
+  if (!mimeType) {
+    return false
+  }
+  const videoEl: any = document.createElement('video')
+  // videoElement.canPlayType() can return 'probably'|'maybe' and '' for unsupported mime types.
+  // It is observed that major browsers like Chrome, Firefox says 'maybe' for widely supported mime types like 'video/mp4' as well.
+  // So checking for both 'probably' and 'maybe' here.
+  return ['probably', 'maybe'].includes(videoEl.canPlayType(mimeType))
+}
+
+
 /**
  * this function filters `list` by `keyword`
  * `list` should be an array of objects and strings
